refactor(native-tools): split directory-detector into modular architecture
Refactored directory-detector.swift (1835 lines) into 8 focused modules:
- Types.swift: Shared types and constants
- CWDDetector.swift: libproc-based CWD detection
- TerminalDetector.swift: Terminal.app and iTerm2 support
- IDEDetector.swift: JetBrains IDEs and VSCode variants
- ProcessTree.swift: Process tree traversal utilities
- FileSearch.swift: fd integration and file listing
- DirectoryDetector.swift: Main public API
- main.swift: CLI entry point

Maintains identical public API and functionality while improving:
- Code organization and maintainability
- Logical separation of concerns
- Easier testing and future modifications
- Reduced cognitive load per file

All commands (detect, detect-with-files, list, check-fd) verified working.
Updated Makefile to compile from new directory structure.
