+++ native/Makefile
@@ -5,7 +5,21 @@ SWIFTC = swiftc
 SWIFT_FLAGS = -O
 FRAMEWORKS = -framework Cocoa -framework ApplicationServices
 OUTPUT_DIR = ../src/native-tools
-SOURCES = window-detector.swift keyboard-simulator.swift text-field-detector.swift directory-detector.swift
+
+# Single-file tools
+SINGLE_FILE_SOURCES = window-detector.swift keyboard-simulator.swift text-field-detector.swift
+
+# Multi-file directory-detector sources (order matters for readability, not compilation)
+DIRECTORY_DETECTOR_SOURCES = \
+	directory-detector/Types.swift \
+	directory-detector/CWDDetector.swift \
+	directory-detector/TerminalDetector.swift \
+	directory-detector/IDEDetector.swift \
+	directory-detector/ProcessTree.swift \
+	directory-detector/FileSearch.swift \
+	directory-detector/DirectoryDetector.swift \
+	directory-detector/main.swift
+
 TARGETS = $(OUTPUT_DIR)/window-detector $(OUTPUT_DIR)/keyboard-simulator $(OUTPUT_DIR)/text-field-detector $(OUTPUT_DIR)/directory-detector
 
 # Bridging header for libproc (used by directory-detector for fast CWD detection)
@@ -27,9 +41,10 @@ $(OUTPUT_DIR)/keyboard-simulator: keyboard-simulator.swift | $(OUTPUT_DIR)
 $(OUTPUT_DIR)/text-field-detector: text-field-detector.swift | $(OUTPUT_DIR)
 	$(SWIFTC) $(SWIFT_FLAGS) $(FRAMEWORKS) -o $@ $<
 
-# directory-detector uses libproc for fast CWD detection (10-50x faster than lsof)
-$(OUTPUT_DIR)/directory-detector: directory-detector.swift libproc-bridge.h | $(OUTPUT_DIR)
-	$(SWIFTC) $(SWIFT_FLAGS) $(FRAMEWORKS) $(BRIDGING_HEADER) -o $@ directory-detector.swift
+# directory-detector: multi-file compilation with libproc bridging header
+# Uses libproc for fast CWD detection (10-50x faster than lsof)
+$(OUTPUT_DIR)/directory-detector: $(DIRECTORY_DETECTOR_SOURCES) libproc-bridge.h | $(OUTPUT_DIR)
+	$(SWIFTC) $(SWIFT_FLAGS) $(FRAMEWORKS) $(BRIDGING_HEADER) -o $@ $(DIRECTORY_DETECTOR_SOURCES)
 
 install: all
 	chmod +x $(TARGETS)
