+++ native/directory-detector/TerminalDetector.swift
@@ -1,0 +1,47 @@
+import Cocoa
+import Foundation
+
+// MARK: - Terminal Detection
+
+extension DirectoryDetector {
+
+    // MARK: - Terminal.app tty detection
+
+    static func getTerminalTty() -> String? {
+        let script = """
+        tell application "Terminal"
+            set W to the front window
+            set T to selected tab of W
+            return tty of T
+        end tell
+        """
+
+        let appleScript = NSAppleScript(source: script)
+        var error: NSDictionary?
+        guard let result = appleScript?.executeAndReturnError(&error) else {
+            return nil
+        }
+
+        return result.stringValue
+    }
+
+    // MARK: - iTerm2 tty detection
+
+    static func getiTerm2Tty() -> String? {
+        let script = """
+        tell application "iTerm2"
+            tell current session of current tab of current window
+                return tty
+            end tell
+        end tell
+        """
+
+        let appleScript = NSAppleScript(source: script)
+        var error: NSDictionary?
+        guard let result = appleScript?.executeAndReturnError(&error) else {
+            return nil
+        }
+
+        return result.stringValue
+    }
+}
